consolidates api-server and auth-server into auth-api
