+++ auth-api/.gitignore
