+++ auth-api/index.js
@@ -1,9 +1,9 @@
 'use strict';
 
 require('dotenv').config();
-const app = require('./src/server.js');
+const server = require('./src/server.js');
 const { db } = require('./src/auth/models');
 
 db.sync().then(() => {
-  app.start(process.env.PORT || 3001);
+  server.start(process.env.PORT || 3001);
 });
