+++ auth-api/src/auth/middleware/bearer.js
