+++ auth-api/src/auth/middleware/logger.js
