+++ auth-api/src/auth/models/index.js
@@ -1,11 +1,12 @@
 'use strict';
 
+const userModel = require('./users.js');
 const { Sequelize, DataTypes } = require('sequelize');
-const clothesModel = require('./clothes/model.js');
-const foodModel = require('./food/model.js');
+const clothesModel = require('./clothes/model.js.js');
+const foodModel = require('./food/model.js.js');
 const Collection = require('./data-collection.js');
 
-const DATABASE_URL = process.env.DATABASE_URL || 'sqlite:memory:';
+const DATABASE_URL = process.env.DATABASE_URL || 'sqlite:memory;';
 
 const sequelize = new Sequelize(DATABASE_URL);
 const food = foodModel(sequelize, DataTypes);
@@ -13,6 +14,7 @@ const clothes = clothesModel(sequelize, DataTypes);
 
 module.exports = {
   db: sequelize,
+  users: userModel(sequelize, DataTypes),
   food: new Collection(food),
   clothes: new Collection(clothes),
 };
