+++ auth-api/src/auth/routes/authRoutes.js
@@ -3,28 +3,28 @@
 const express = require('express');
 const authRouter = express.Router();
 
-const { users } = require('./models');
-const basicAuth = require('./middleware/basic.js')
-const bearerAuth = require('./middleware/bearer.js')
-const permissions = require('./middleware/acl.js')
+const { users } = require('../models');
+const basicAuth = require('../middleware/basic.js');
+const bearerAuth = require('../middleware/bearer.js');
+const permissions = require('../middleware/acl.js');
 
 authRouter.post('/signup', async (req, res, next) => {
   try {
     let userRecord = await users.create(req.body);
     const output = {
       user: userRecord,
-      token: userRecord.token
+      token: userRecord.token,
     };
     res.status(201).json(output);
   } catch (e) {
-    next(e.message)
+    next(e.message);
   }
 });
 
 authRouter.post('/signin', basicAuth, (req, res, next) => {
   const user = {
     user: req.user,
-    token: req.user.token
+    token: req.user.token,
   };
   res.status(200).json(user);
 });
@@ -36,7 +36,7 @@ authRouter.get('/users', bearerAuth, permissions('delete'), async (req, res, nex
 });
 
 authRouter.get('/secret', bearerAuth, async (req, res, next) => {
-  res.status(200).send('Welcome to the secret area')
+  res.status(200).send('Welcome to the secret area');
 });
 
 module.exports = authRouter;
