+++ auth-api/src/error-handlers/404.js
