+++ auth-api/src/error-handlers/500.js
@@ -16,7 +16,7 @@ module.exports = function (err, req, res, next) {
 
   const errorObject = {
     status: 500,
-    message: error
+    message: error,
   };
   res.status(500).json(errorObject);
 };
