+++ auth-api/src/server.js
@@ -1,5 +1,7 @@
 'use strict';
 
+// Note: Includes both auth and api server.js content (from original starter code)
+
 // 3rd Party Resources
 const express = require('express');
 const cors = require('cors');
@@ -8,7 +10,13 @@ const morgan = require('morgan');
 // Esoteric Resources
 const errorHandler = require('./error-handlers/500.js');
 const notFound = require('./error-handlers/404.js');
-const authRoutes = require('./auth/routes.js');
+const logger = require('./middleware/logger.js.js');
+
+// Auth routes
+const authRoutes = require('./auth/routes/authRoutes.js');
+
+// API routes
+const v1Routes = require('./routes/v1.js.js');
 
 // Prepare the express app
 const app = express();
@@ -16,22 +24,24 @@ const app = express();
 // App Level MW
 app.use(cors());
 app.use(morgan('dev'));
+app.use(logger);
 
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 
+
 // Routes
 app.use(authRoutes);
+app.use('/api/v1', v1Routes);
 
 // Catchalls
-app.use(notFound);
+app.use('*', notFound);
 app.use(errorHandler);
 
 module.exports = {
   server: app,
-  start: (port) => {
-    app.listen(port, () => {
-      console.log(`Server Up on ${port}`);
-    });
+  start: port => {
+    if (!port) { throw new Error('Missing Port'); }
+    app.listen(port, () => console.log(`Listening on ${port}`));
   },
 };
